+++ CHANGELOG.md
@@ -5,7 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-
+### Added
+- Allow `s.str`'s `regex` kwarg to be a pre-existing pattern returned from
+  `re.compile` (#36)
 
 ## [v0.2.1] - 2019-10-28
 ### Fixed